move components to search view (#739)
