+++ packages/gatsby-theme-store/src/templates/search.browser.tsx
@@ -5,8 +5,8 @@ import type { PageProps as GatsbyPageProps } from 'gatsby'
 
 import HybridWrapper from '../components/HybridWrapper'
 import Layout from '../components/Layout'
-import SearchView from '../components/SearchPage'
-import AboveTheFoldPreview from '../components/SearchPage/AboveTheFoldPreview'
+import SearchView from '../views/search'
+import AboveTheFoldPreview from '../views/search/AboveTheFoldPreview'
 import { useQuery } from '../sdk/graphql/useQuery'
 import { useQueryVariablesFromSearchParams } from '../sdk/search/converter/useQueryVariablesFromSearchParams'
 import { useSearchParamsFromUrl } from '../sdk/search/converter/useSearchParamsFromURL'
