+++ packages/gatsby-theme-store/src/views/home/AboveTheFold.tsx
