+++ packages/gatsby-theme-store/src/views/home/BelowTheFold.tsx
