+++ packages/gatsby-theme-store/src/views/home/BelowTheFoldPreview.tsx
