+++ packages/gatsby-theme-store/src/views/home/SEO/index.tsx
