+++ packages/gatsby-theme-store/src/views/home/SEO/useMetadata.ts
