+++ packages/gatsby-theme-store/src/views/search/BelowTheFold.tsx
