+++ packages/gatsby-theme-store/src/views/search/BelowTheFoldPreview.tsx
