+++ packages/gatsby-theme-store/src/views/search/SEO/index.tsx
