+++ packages/gatsby-theme-store/src/views/search/SEO/useMetadata.ts
