+++ packages/gatsby-theme-store/src/views/search/index.tsx
